+++ app/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Kakaosearch">
-        <activity android:name="io.github.beomjo.search.MainActivity">
+        <activity android:name="io.github.beomjo.search.ui.activity.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
