+++ app/src/main/kotlin/io/github/beomjo/search/ui/fragment/SearchFragment.kt
@@ -14,32 +14,21 @@
  * limitations under the License.
  */
 
-package io.github.beomjo.search.ui
+package io.github.beomjo.search.ui.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.findNavController
-import io.github.beomjo.search.databinding.FragmentHomeBinding
+import io.github.beomjo.search.R
 
-class HomeFragment : Fragment() {
+class SearchFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View {
-        val binding = FragmentHomeBinding.inflate(layoutInflater)
-        setView(binding)
-        return binding.root
-    }
-
-    private fun setView(binding: FragmentHomeBinding) {
-        binding.text.setOnClickListener {
-            HomeFragmentDirections.actionHomeDestToDetailDest().let {
-                findNavController().navigate(it)
-            }
-        }
+    ): View? {
+        return inflater.inflate(R.layout.fragment_search, container, false)
     }
 }
