+++ app/src/main/res/layout/fragment_detail.xml
@@ -15,8 +15,10 @@
     -->
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    tools:context="io.github.beomjo.search.ui.fragment.DetailFragment">
 
     <TextView
         android:layout_width="0dp"
