+++ app/src/main/res/layout/fragment_home.xml
@@ -18,7 +18,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="io.github.beomjo.search.ui.HomeFragment">
+    tools:context="io.github.beomjo.search.ui.fragment.HomeFragment">
 
     <TextView
         android:id="@+id/text"
