+++ app/src/main/res/layout/fragment_search.xml
@@ -18,7 +18,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="io.github.beomjo.search.ui.SearchFragment">
+    tools:context="io.github.beomjo.search.ui.fragment.SearchFragment">
 
     <TextView
         android:layout_width="0dp"
